Move Sca buildinfo logic to technologies package
