+++ sca/bom/buildinfo/technologies/cocoapods/cocoapods_test.go
