+++ sca/bom/buildinfo/technologies/cocoapods/podcommand.go
