+++ sca/bom/buildinfo/technologies/go/golang.go
