+++ sca/bom/buildinfo/technologies/go/golang_test.go
