+++ sca/bom/buildinfo/technologies/java/deptreemanager.go
