+++ sca/bom/buildinfo/technologies/java/deptreemanager_test.go
