+++ sca/bom/buildinfo/technologies/java/gradle.go
