+++ sca/bom/buildinfo/technologies/java/gradle_test.go
