+++ sca/bom/buildinfo/technologies/java/mvn_test.go
