+++ sca/bom/buildinfo/technologies/java/resources/settings.xml
