+++ sca/bom/buildinfo/technologies/npm/npm.go
