+++ sca/bom/buildinfo/technologies/npm/npm_test.go
