+++ sca/bom/buildinfo/technologies/nuget/nuget_test.go
