+++ sca/bom/buildinfo/technologies/pnpm/pnpm.go
