+++ sca/bom/buildinfo/technologies/pnpm/pnpm_test.go
