+++ sca/bom/buildinfo/technologies/python/python.go
