+++ sca/bom/buildinfo/technologies/python/python_test.go
