+++ sca/bom/buildinfo/technologies/swift/swift.go
