+++ sca/bom/buildinfo/technologies/swift/swift_test.go
