+++ sca/bom/buildinfo/technologies/swift/swiftcommand.go
