+++ sca/bom/buildinfo/technologies/yarn/yarn.go
