+++ sca/bom/buildinfo/technologies/yarn/yarn_test.go
